javascript restructuring
