+++ fermo_gui/static/js/dynamic_tables.js
@@ -1,0 +1,339 @@
+// Functions to dynamically update tables after feature or sample selection //
+// TODO: remove general feature data after sample switch
+export function updateTableWithFeatureData(fId, sampleData) {
+    document.getElementById('featureIdCell').textContent = sampleData.featureId[fId];
+    document.getElementById('precMzCell').textContent = sampleData.precMz[fId];
+    document.getElementById('retTimeCell').textContent = sampleData.retTime[fId];
+    document.getElementById('relIntCell').textContent = sampleData.relInt[fId];
+    document.getElementById('absIntCell').textContent = sampleData.absInt[fId];
+    document.getElementById('NovScore').textContent = sampleData.novScore[fId];
+    document.getElementById('BlankAs').textContent = sampleData.blankAs[fId];
+}
+
+// TODO: test multiple groups
+export function updateTableWithGroupData(groupData){
+    var featureData = Object.entries(groupData);
+    if (Object.keys(featureData).length === 0) {
+        document.getElementById('feature-general-info').textContent = 'No group data available for this feature.'
+        Plotly.purge('heatmap-container');
+    } else {
+        document.getElementById('feature-general-info').textContent = 'The feature is found in the following groups:'
+        for (const [key, value] of featureData) {
+            createHeatmap(value, key);
+        }
+    }
+}
+
+export function updateTableWithSampleData(sampleData) {
+    let tableBody = document.getElementById("sampleCell");
+    tableBody.innerHTML = "";
+    let dataArray = sampleData;
+    dataArray.forEach(item => {
+        let row = document.createElement("tr");
+        let sIdCell = document.createElement("td");
+        sIdCell.textContent = item.s_id;
+        row.appendChild(sIdCell);
+        let valueCell = document.createElement("td");
+        valueCell.textContent = item.value;
+        row.appendChild(valueCell);
+        tableBody.appendChild(row);
+    });
+}
+
+// TODO: remove feature data after sample switch
+export function updateTableWithAnnotationData(annotations, sample) {
+    // Define headers and columns for each section
+    let matchHeaders = ["Match id", "Score", "More info"];
+    let matchColumns = ["id", "score"];
+    let matchExtraColumns = [
+        { title: "Algorithm", field: "algorithm" },
+        { title: "Mz", field: "mz" },
+        { title: "Difference in mz", field: "diff_mz" },
+        { title: "SMILES", field: "smiles" },
+        { title: "Link to MIBiG", field: "id" }
+    ];
+
+    let lossHeaders = ["Loss id", "Difference in ppm", "More info"];
+    let lossColumns = ["id", "diff_ppm"];
+    let lossExtraColumns = [
+        { title: "Detected loss", field: "det_loss" },
+        { title: "Expected loss", field: "exp_loss" },
+        { title: "Fragment mz", field: "mz_frag" }
+    ];
+
+    let fragmentHeaders = ["Fragment id", "Difference in ppm", "More info"];
+    let fragmentColumns = ["id", "diff_ppm"];
+    let fragmentExtraColumns = [
+        { title: "Detected fragment", field: "frag_det" },
+        { title: "Expected fragment", field: "frag_ex" }
+    ];
+
+    let adductHeaders = ["Adduct type", "Difference in ppm", "More info"];
+    let adductColumns = ["adduct_type", "diff_ppm"];
+    let adductExtraColumns = [
+        { title: "Partner adduct", field: "partner_adduct" },
+        { title: "Partner mz", field: "partner_mz" },
+        { title: "Partner id", field: "partner_id" },
+        { title: "difference in ppm", field: "diff_ppm" }
+    ];
+
+    // create tables for each section
+    let isAnyDataPresent = false;
+
+    isAnyDataPresent = createAnnotationTable("matchTable", annotations.matches, matchHeaders,
+    matchColumns, matchExtraColumns, "match") || isAnyDataPresent;
+    isAnyDataPresent = createAnnotationTable("lossesTable", annotations.losses, lossHeaders,
+    lossColumns, lossExtraColumns, "loss") || isAnyDataPresent;
+    isAnyDataPresent = createAnnotationTable("fragmentTable", annotations.fragments, fragmentHeaders,
+    fragmentColumns, fragmentExtraColumns, "fragment") || isAnyDataPresent;
+    isAnyDataPresent = createAnnotationTable("adductTable", annotations.adducts, adductHeaders,
+    adductColumns, adductExtraColumns, "adduct") || isAnyDataPresent;
+
+    let annMessage = document.getElementById("feature-annotation");
+    if (!isAnyDataPresent) {
+        annMessage.innerHTML = "No annotation data found for this feature.";
+    }
+}
+
+function createAnnotationTable(sectionId, annotations, headers, columns, extraColumns, tableId) {
+    let tableBody = document.getElementById(sectionId);
+    tableBody.innerHTML = "";
+
+    let annMessage = document.getElementById("feature-annotation");
+    annMessage.innerHTML = "";
+
+    if (annotations.length > 0) {
+        let tHead = createTableHeader(headers);
+        tableBody.appendChild(tHead);
+
+        let tBody = createTableBody(annotations, columns, extraColumns, tableId);
+        tableBody.appendChild(tBody);
+
+        tableBody.style.marginBottom = "1rem";
+        return true;
+    } else {
+        tableBody.style.marginBottom = "0";
+        return false;
+    }
+}
+
+function createTableHeader(headers) {
+    let tHead = document.createElement("thead");
+    let rowHead = document.createElement("tr");
+
+    headers.forEach(header => {
+        let th = document.createElement("th");
+        th.textContent = header;
+        rowHead.appendChild(th);
+    });
+
+    tHead.appendChild(rowHead);
+    return tHead;
+}
+
+function createTableRow(data, columns, rowId, tableId) {
+    let row = document.createElement("tr");
+
+    columns.forEach(column => {
+        let td = document.createElement("td");
+        let cellData = data[column];
+        if (typeof cellData === 'string' && cellData.includes("|")) {
+            td.textContent = cellData.split("|")[0];
+        } else {
+            td.textContent = cellData !== undefined ? cellData : '';
+        }
+        row.appendChild(td);
+    });
+
+    // Add the collapse button for expanding extra info
+    let expandCell = document.createElement("td");
+    expandCell.className = "text-center";
+    let expandButton = document.createElement("button");
+    expandButton.className = "accordion-button-info collapsed";
+    expandButton.type = "button";
+    expandButton.setAttribute("data-bs-toggle", "collapse");
+    expandButton.setAttribute("data-bs-target", `#collapse-${tableId}-${rowId}`);
+    expandButton.setAttribute("aria-expanded", "false");
+    expandButton.setAttribute("aria-controls", `collapse-${tableId}-${rowId}`);
+
+    expandCell.appendChild(expandButton);
+    row.appendChild(expandCell);
+
+    return row;
+}
+
+function createTableBody(dataArray, columns, extraColumns, tableId) {
+    let tBody = document.createElement("tbody");
+
+    dataArray.forEach((data, index) => {
+        let row = createTableRow(data, columns, index, tableId);
+        tBody.appendChild(row);
+
+        // Create the extra info row
+        let extraInfoRow = document.createElement("tr");
+        let extraInfoCell = document.createElement("td");
+        extraInfoCell.colSpan = columns.length + 1;
+        extraInfoCell.style.padding = "0px";
+
+        let collapseDiv = document.createElement("div");
+        collapseDiv.id = `collapse-${tableId}-${index}`;
+        collapseDiv.className = "collapse";
+
+        let { extraInfo } = createExtraInfoRow(data, extraColumns, columns.length + 1);
+        collapseDiv.appendChild(extraInfo);
+
+        extraInfoCell.appendChild(collapseDiv);
+        extraInfoRow.appendChild(extraInfoCell);
+
+        tBody.appendChild(extraInfoRow);
+    });
+
+    return tBody;
+}
+
+function clickToCopy(data, displayText) {
+    const span = document.createElement("span");
+    span.textContent = displayText;
+    span.style.cursor = "pointer";
+    span.style.color = "rgba(13, 110, 253)";
+
+    span.onclick = function() {
+        const tempInput = document.createElement("input");
+        tempInput.value = data;
+        document.body.appendChild(tempInput);
+        tempInput.select();
+        document.execCommand("copy");
+        document.body.removeChild(tempInput);
+
+        alert("Copied to clipboard: \n" + data);
+    };
+
+    return span;
+}
+
+function createExtraInfoRow(data, extraColumns, colspan) {
+    let extraInfo = document.createElement("div");
+    extraInfo.style.padding = "10px";
+    extraInfo.style.backgroundColor = "#f9f9f9";
+    extraInfo.style.fontSize = "0.8em";
+
+    let infoTable = document.createElement("table");
+    infoTable.classList.add("fixed-width-table-extra");
+
+    extraColumns.forEach(column => {
+        let cellData = data[column.field];
+        let infoRow = document.createElement("tr");
+
+        let titleCell = document.createElement("td");
+        titleCell.textContent = column.title + ":";
+        titleCell.style.fontWeight = "bold";
+        infoRow.appendChild(titleCell);
+
+        let valueCell = document.createElement("td");
+        if (column.title === "Link to MIBiG") {
+            if (typeof cellData === 'string' && cellData.includes("|")) {
+                let mibigLink = document.createElement('a');
+                mibigLink.href = `https://mibig.secondarymetabolites.org/repository/${cellData.split("|")[1]}`;
+                mibigLink.textContent = cellData.split("|")[1];
+                mibigLink.target = "_blank";
+                valueCell.appendChild(mibigLink);
+            } else {
+                return;
+            }
+        } else {
+            valueCell.textContent = cellData !== undefined ? cellData : '';
+        }
+
+        if (column.title === "SMILES") {
+            let displayText = "Copy SMILES";
+            let copySpan = clickToCopy(cellData, displayText);
+            valueCell.innerHTML = '';
+            valueCell.appendChild(copySpan);
+        }
+
+        infoRow.appendChild(valueCell);
+        infoTable.appendChild(infoRow);
+    });
+
+    extraInfo.appendChild(infoTable);
+
+    return { extraInfo };
+}
+
+// Heatmap visualization function //
+function createHeatmap(data, groupName) {
+	const groupValues = [...new Set([...data.map(item => item.group1), ...data.map(item => item.group2)])].sort();
+	const matrix = groupValues.map(() => Array(groupValues.length).fill(0));
+
+	// Fill the matrix with factor values
+	data.forEach(item => {
+		const rowIndex = groupValues.indexOf(item.group1);
+		const colIndex = groupValues.indexOf(item.group2);
+		matrix[rowIndex][colIndex] = item.factor;
+		matrix[colIndex][rowIndex] = item.factor;
+	});
+	const reversedMatrix = matrix.reverse();
+	const reversedGroupValues = [...groupValues].reverse();
+
+	// Create value annotations to be shown within the heatmap
+	const annotations = [];
+	for (let i = 0; i < reversedGroupValues.length; i++) {
+		for (let j = 0; j < groupValues.length; j++) {
+			if (reversedMatrix[i][j] !== null) {
+				annotations.push({
+					x: groupValues[j],
+					y: reversedGroupValues[i],
+					text: reversedMatrix[i][j].toFixed(2),
+					showarrow: false,
+					font: { size: 7 }
+				});
+			}
+		}
+	}
+
+	// Define colorscale
+	const colorscaleValue = [
+		[0, '#f5f5f5'],
+		[0.01, '#dbf3ff'],
+		[1, '#de4b4b']
+	];
+
+	// Create a new Plotly figure
+	const trace = {
+		z: reversedMatrix,
+		x: groupValues,
+		y: reversedGroupValues,
+		type: 'heatmap',
+		colorscale: colorscaleValue,
+		showscale: false
+	};
+
+	// Define layout options
+	const layout = {
+		title: {
+			text: `Fold-differences of <br><b>${groupName}</b>`,
+			font: { size: 12 },
+			x: 0.5,
+			xanchor: 'center',
+		},
+		width: 250,
+		height: 250,
+		xaxis: {
+			ticks: '',
+			ticksuffix: ' ',
+			autosize: false,
+			font: { size: 7 },
+		},
+		yaxis: {
+			ticks: '',
+			ticksuffix: ' ',
+			autosize: false,
+			font: { size: 7 },
+		},
+		annotations: annotations,
+		margin: { l: 50, r: 50, t: 50, b: 50 },
+	};
+
+	// Create the heatmap
+	Plotly.newPlot('heatmap-container', [trace], layout);
+}
