+++ fermo_gui/templates/dashboard.html
@@ -216,48 +216,6 @@
                                 </div>
                             </div>
                         </div>
-                        <div class="accordion accordion-item" draggable="false" id="quantFeature">
-                            <h2 class="accordion-header" id="quantTableGroups">
-                                <button class="accordion-button collapsed"
-                                        type="button"
-                                        data-bs-toggle="collapse"
-                                        data-bs-target="#featureQuant"
-                                        aria-expanded="false"
-                                        aria-controls="featureTable">
-                                    Quant-data information
-                                </button>
-                            </h2>
-                            <div id="featureQuant"
-                                 class="accordion-collapse collapse"
-                                 aria-labelledby="featureTableGeneral">
-                                <div class="accordion-body overflow-auto">
-                                    <table class="table table-striped table-hover">
-                                        <tbody>
-                                            <tr>
-                                                <td style="text-align: left;">QuantData-associated</td>
-                                                <td id="quantAsCell" style="text-align: left;"></td>
-                                            </tr>
-                                            <tr>
-                                                <td style="text-align: left;">QuantData-trend</td>
-                                                <td id="quantTrendCell" style="text-align: left;"></td>
-                                            </tr>
-                                            <tr>
-                                                <td style="text-align: left;">QuantData per sample</td>
-                                                <td id="quantGroupCell" style="text-align: left;">Click on any feature in the main chromatogram overview</td>
-                                            </tr>
-                                            <tr>
-                                                <td style="text-align: left;">Original QuantData</td>
-                                                <td id="quantCell" style="text-align: left;"></td>
-                                            </tr>
-                                            <tr>
-                                                <td style="text-align: left;">Putative adducts</td>
-                                                <td id="AdductsCell" style="text-align: left;"></td>
-                                            </tr>
-                                        </tbody>
-                                    </table>
-                                </div>
-                            </div>
-                        </div>
                     </div>
                 </div>
             </div>
@@ -319,21 +277,43 @@
             </div>
         </div>
 
-        <div class="dropzone">
-            <div class="accordion accordion-item" draggable="true" id="cyto">
-                <h2 class="accordion-header" id="cytoscapeHeader">
-                    <button class="accordion-button collapsed" type="button" data-bs-toggle="collapse" data-bs-target="#cytoscape" aria-expanded="true" aria-controls="Cytoscape">
-                        Cytoscape
-                    </button>
-                </h2>
-                <div class="accordion-collapse collapse" id="cytoscape" aria-labelledby="CytoscapeHeader">
-                    <div class="accordion-body overflow-auto">
-                        <h6 class="fw-bold"> Cytoscape view - spectral similarity networking</h6>
-                        {% if cytoscape_message %}
-                            <p class="text-danger">{{ cytoscape_message }}</p>
-                        {% endif %}
-                        <div id="cy"></div>
-                        <div class="d-inline-flex mt-2">
+        <div class="row">
+            <div class="col-xs-12 col-sm-12 col-md-6">
+                <div class="dropzone">
+                    <div class="accordion accordion-item" draggable="true" id="cyto">
+                        <h2 class="accordion-header" id="cytoscapeHeader">
+                            <button class="accordion-button collapsed" type="button" data-bs-toggle="collapse" data-bs-target="#cytoscape" aria-expanded="true" aria-controls="Cytoscape">
+                                Spectral similarity networking
+                            </button>
+                        </h2>
+                        <div class="accordion-collapse collapse" id="cytoscape" aria-labelledby="CytoscapeHeader">
+                            <div class="accordion-body overflow-auto">
+                                <p id="activeFeature">Select any feature in the main chromatogram to visualize its network.</p>
+                                <div id="cy"></div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+            <div class="col-xs-12 col-sm-12 col-md-6">
+                <div class="dropzone">
+                    <div class="accordion accordion-item" draggable="true" id="dashboardSettings">
+                        <h2 class="accordion-header" id="settingsHeader">
+                            <button class="accordion-button collapsed"
+                                    type="button"
+                                    data-bs-toggle="collapse"
+                                    data-bs-target="#settings"
+                                    aria-expanded="false"
+                                    aria-controls="settings">
+                                Dashboard settings
+                            </button>
+                        </h2>
+                        <div id="settings"
+                             class="accordion-collapse collapse"
+                             aria-labelledby="filtersHeader">
+                            <div class="accordion-body overflow-auto">
+
+                            </div>
                         </div>
                     </div>
                 </div>
@@ -363,8 +343,7 @@
 
 
 
-<script type="text/javascript" src="{{ url_for('static', filename='js/dashboard.js') }}"></script>
-<script type="text/javascript" src="{{ url_for('static', filename='js/chromatogram.js') }}"></script>
+<script type="module" src="{{ url_for('static', filename='js/dashboard.js') }}"></script>
 <script type="text/javascript" src="{{ url_for('static', filename='js/plotly-2.18.2.min.js') }}"></script>
 <script>
     var statsChromatogram = {{ data.stats_chromatogram | tojson | safe }};
